Remove `metabase-types/types/Parameters`
