+++ frontend/src/metabase-lib/parameters/utils/cards.ts
@@ -1,5 +1,4 @@
-import { Parameter } from "metabase-types/api";
-import { ParameterTarget } from "metabase-types/types/Parameter";
+import { Parameter, ParameterTarget } from "metabase-types/api";
 import { Card } from "metabase-types/types/Card";
 import {
   ParameterWithTarget,
