+++ frontend/src/metabase-types/api/dashboard.ts
@@ -1,10 +1,10 @@
-import { Parameter } from "metabase-types/api/parameters";
 import type {
-  ParameterTarget,
+  Parameter,
   ParameterId,
-} from "metabase-types/types/Parameter";
+  ParameterTarget,
+} from "metabase-types/api";
 
-import { ActionDashboardCard } from "./actions";
+import type { ActionDashboardCard } from "./actions";
 import type { SearchModelType } from "./search";
 import type { Card, CardId, CardDisplayType } from "./card";
 import type { Dataset } from "./dataset";
